==> data_prcssng/01_etl_movies_subset_limpiar.ipynb <==
# 1.1. ELT - Fase 1: Limpieza de datos segun requerimientos minimos de Henry

En esta seccion, limpiamos el dataset movies_dataset.csv segun los requerimientos minimos de Henry; son 6 instrucciones en total:
1. Eliminar las columnas que no serán utilizadas, video,imdb_id,adult,original_title,poster_path y homepage.
2. De haber fechas, deberán tener el formato AAAA-mm-dd
3. Los valores nulos del campo release date deben eliminarse
4. Deberán crear la columna release_year donde extraerán el año de la fecha de estreno.
5. Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0
6. Crear la columna con el retorno de inversión, llamada return con los campos revenue y budget, dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0

# requisitos
import os
import pandas as pd
import numpy as np

data_movies = pd.read_csv(
    os.path.join('1_data','movies_dataset.zip'),
    compression='zip',low_memory=False).convert_dtypes()
data_movies.rename(columns = {'id':'pelicula_id'}, inplace = True) # definimos nuestro primary key

## Instruccion 1

Eliminar las columnas que no serán utilizadas: video, imdb_id, adult, original_title, poster_path y homepage.

columns_excluidas_henry=[ # Instruccion: Eliminar las columnas que no serán utilizadas, video,imdb_id,adult,original_title,poster_path y homepage
    'video',
    'imdb_id',
    'adult',
    'original_title',
    'poster_path',
    'homepage']
columns_excluidas_crtrprsnl=['status','tagline'] # criterio personal: no relevantes para modelo ML
columns_desanidar=[
    'belongs_to_collection',
    'genres',
    'production_companies',
    'production_countries',
    'spoken_languages']
data_movies_subset=data_movies.drop(list(columns_excluidas_henry+columns_excluidas_crtrprsnl+columns_desanidar), axis=1)
data_movies_subset.info()
data_movies_subset.head(2)

Esperamos que las siguientes columnas sean numericas: pelicula_id
pero no lo es.

#data_movies_subset['pelicula_id']=pd.to_numeric(data_movies_subset['pelicula_id']) # ValueError: Unable to parse string "1997-08-20" at position 19730
# inspeccionamos visualmente el error arojado en la posicion 19730 de pelicula_id
data_movies_subset.iloc[19730]['pelicula_id'] # algunas filas puede contener valores que no corresponden al valor esperado

# identificamos las filas que contienen datos tipo fecha en la columna de nuestro primary key: pelicula_id
def check_type(x):
     try: 
        return type(eval(x)) 
     except Exception as e:
        return type(x)

data_movies_subset.loc[data_movies_subset['pelicula_id'].map(check_type)!=int]

# removemos las filas que contienen fecha en la columna de primary key
data_movies_subset.drop(
    data_movies_subset.loc[data_movies_subset['pelicula_id'].map(check_type)!=int].index,
    inplace=True)
print("Numero de filas que no contienen datos tipo fecha en la columna ['pelicula_id']:",
    data_movies_subset.loc[data_movies_subset['pelicula_id'].map(check_type)!=int].shape[0]
)

## Instruccion 2
De haber fechas, deberán tener el formato AAAA-mm-dd

columns_numeric=[
    'pelicula_id',
    'budget',
    'revenue',
    'popularity',
    'vote_average',
    'vote_count',
    'runtime']

columns_date=['release_date']

data_movies_subset[columns_numeric]=data_movies_subset[columns_numeric].apply(pd.to_numeric) # convertimos a numerica
data_movies_subset[columns_date]=data_movies_subset[columns_date].apply(pd.to_datetime) # convertimos a fecha 

data_movies_subset_limpia=data_movies_subset.convert_dtypes() # volvemos a evaluar el df
data_movies_subset_limpia.info()
data_movies_subset_limpia.head(3)

## Instruccion 3
Los valores nulos del campo release date deben eliminarse

# instruccion 1: Los valores nulos del campo release date deben eliminarse
data_movies_subset_limpia['release_date'].isna().any()
print("Cantidad de valores nulos en ['release_date']:",
    data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].shape[0])
data_movies_subset_limpia['release_date'].isna()
rows_release_date_na = data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].index
data_movies_subset_limpia.drop(rows_release_date_na, inplace=True)
print("Cantidad de valores nulos en ['release_date'] despues del drop:",
    data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].shape[0])

## Instruccion 4
Deberán crear la columna release_year donde extraerán el año de la fecha de estreno.

data_movies_subset_limpia['release_year'] = data_movies_subset_limpia['release_date'].dt.year

## Instruccion 5

Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0

# Cantidad de valores nulos en ['budget','revenue']
data_movies_subset_limpia[['budget','revenue']].isna().sum()

data_movies_subset_limpia[['budget','revenue']] = data_movies_subset_limpia[['budget','revenue']].fillna(0)
data_movies_subset_limpia[['budget','revenue']].isna().sum()

## Instruccion 6
Crear la columna con el retorno de inversión, llamada return con los campos revenue y budget, dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0

data_movies_subset_limpia['return']=data_movies_subset_limpia['revenue']/data_movies_subset_limpia['budget'] # nueva columna return
data_movies_subset_limpia['return']=np.nan_to_num(data_movies_subset_limpia['return']) # convertimos a np.num para facilar el reemplazo a zeros
data_movies_subset_limpia['return'].fillna(0, inplace=True) # reemplazamos por zero cuando no hay datos para el caso de NaN
data_movies_subset_limpia['return'].replace([np.inf, -np.inf], 0, inplace=True) # reemplazamos por zero cuando no hay datos para el caso de inf
print("Hay valores nulos en cualquiera de ['revenue','budget','return']:",
    any(data_movies_subset_limpia[['revenue','budget','return']].isna().any()))


# Removemos duplicados
data_movies_subset_limpia.drop_duplicates(inplace=True) # no funciona; deben existir filas que no son exactamente iguales.
print("Numero de peliculas unicas:",data_movies_subset_limpia['pelicula_id'].nunique(),"\n")
print("Numero de registros (filas):",data_movies_subset_limpia.shape[0],"\n")
print("Numero de peliculas con informacion inconsiste:",data_movies_subset_limpia.shape[0]-data_movies_subset_limpia['pelicula_id'].nunique(),"\n")

# para una misma pelicula existen valores distintos.
data_movies_subset_inconsistent=data_movies_subset_limpia.loc[data_movies_subset_limpia.duplicated(subset=['pelicula_id'], keep=False)]
print("Numero de filas con informacion duplicada en algunas columnas:",
    data_movies_subset_inconsistent.shape[0])
print("Numero de peliculas unicas con informacion duplicadas en algunas columnas:",
    data_movies_subset_inconsistent['pelicula_id'].nunique())
data_movies_subset_inconsistent.sort_values(['pelicula_id']).head(4)

# Solucionamos tomando la primera entrada para cada pelicula, y descartamos el resto.
data_movies_subset_limpia_sindups=data_movies_subset_limpia.loc[~data_movies_subset_limpia.duplicated(subset=['pelicula_id'], keep='first')]
print("Numero de peliculas unicas:",data_movies_subset_limpia_sindups['pelicula_id'].nunique(),"\n")
print("Numero de registros (filas):",data_movies_subset_limpia_sindups.shape[0],"\n")
print("Numero de peliculas con informacion inconsiste:",
      data_movies_subset_limpia_sindups.shape[0]-data_movies_subset_limpia_sindups['pelicula_id'].nunique(),"\n")


## Finalizamos
Guardamos el dataset limpio

# Nuestro dataset final
print("Universo de peliculas con datos unicos:",data_movies_subset_limpia_sindups['pelicula_id'].nunique(),"\n")
data_movies_subset_limpia_sindups.to_csv(os.path.join("2_pipeline","data_movies_subset_limpia.csv"))
data_movies_subset_limpia_sindups.info()
data_movies_subset_limpia_sindups.head(2)
