==> data_prcssng/02_etl_movies_desanidar_v3.ipynb <==
# 1.2. ETL - Fase 2: Desanidar y normalizar movies

En esta seccion, desanidamos y normalizamos las columnas de `movies_dataset.csv` que lo requieren

import os
import zipfile
import pandas as pd
import json
from ast import literal_eval

data_movies = pd.read_csv(os.path.join('1_data','movies_dataset.zip'),compression='zip',low_memory=False).convert_dtypes()
data_movies.rename(columns = {'id':'pelicula_id'}, inplace = True)
data_movies.info()

# inspeccionamos visualmente cuales solumnas requieres ser desanidadas y normalizadas
pd.set_option('display.max_columns', None) # para poder ver todas las columnas
data_movies.head(2)

# nos concentramos en: belongs_to_collection
data_movies_collection=pd.DataFrame(data_movies[['belongs_to_collection','pelicula_id']]).dropna()
data_movies_collection.head(2)

data_movies_collection['belongs_to_collection']=data_movies_collection['belongs_to_collection'].apply(literal_eval)
data_movies_collection.head(2)

# tercero se normaliza la columna
data_movies_collection=data_movies_collection.set_index('pelicula_id')
data_movies_collection_unnested = pd.json_normalize(data_movies_collection['belongs_to_collection']).set_index(data_movies_collection.index)
data_movies_collection_unnested.head(2)

# finalizamos el df
data_movies_collection_unnested.rename(
    columns = {'id':'franquicia_id','name':'franquicia'},
    inplace = True)
data_movies_collection_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_movies_collection_unnested.head(2)

# nos concentramos en: production_companies
data_movies_companies=pd.DataFrame(data_movies[['production_companies','pelicula_id']]).dropna()
data_movies_companies.head(2)

data_movies_companies['production_companies']=data_movies_companies['production_companies'].apply(literal_eval)
data_movies_companies.head(2)

# segundo, expandimos la columna para que queden diccionarios anidados (deja de ser una lista de diccionarios)
data_movies_companies_exploded = data_movies_companies.explode('production_companies', ignore_index=True)
data_movies_companies_exploded.head(2)

# tercero se normaliza la columna
data_movies_companies_exploded=data_movies_companies_exploded.set_index('pelicula_id')
data_movies_companies_exploded_unnested = pd.json_normalize(data_movies_companies_exploded['production_companies']).set_index(data_movies_companies_exploded.index)
data_movies_companies_exploded_unnested.head(2)

# se finaliza el df
data_movies_companies_exploded_unnested.rename(
    columns = {'id':'productora_id','name':'productora'},
    inplace = True)
data_movies_companies_exploded_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_movies_companies_exploded_unnested.head(2)

# nos concentramos en:  genres 
data_movies_genres=pd.DataFrame(data_movies[['genres','pelicula_id']]).dropna()
data_movies_genres.head(2)

# replicamos los pasos para el caso de  lista de diccionario
data_movies_genres['genres']=data_movies_genres['genres'].apply(literal_eval)
data_movies_genres_exploded = data_movies_genres.explode('genres', ignore_index=True)
data_movies_genres_exploded=data_movies_genres_exploded.set_index('pelicula_id')
data_movies_genres_exploded_unnested = pd.json_normalize(data_movies_genres_exploded['genres']).set_index(data_movies_genres_exploded.index)
data_movies_genres_exploded_unnested.rename(
    columns = {'id':'genres_id','name':'genres'},
    inplace = True)
data_movies_genres_exploded_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_movies_genres_exploded_unnested.head(2)

data_movies_countries = pd.DataFrame(data_movies[['production_countries', 'pelicula_id']]).dropna()
data_movies_countries['production_countries'] = data_movies_countries['production_countries'].apply(literal_eval)
data_movies_countries_exploded = data_movies_countries.explode('production_countries')
data_movies_countries_exploded_unnested = pd.json_normalize(data_movies_countries_exploded['production_countries']).set_index(data_movies_countries_exploded.pelicula_id)
data_movies_countries_exploded_unnested.reset_index(inplace=True)
data_movies_countries_exploded_unnested.rename(
    columns={'iso_3166_1':'pais_isocode','name' : 'pais_name'},
    inplace=True
    )
data_movies_countries_exploded_unnested.head(2)

data_movies_languages = pd.DataFrame(data_movies[['spoken_languages', 'pelicula_id']]).dropna()
data_movies_languages['spoken_languages']= data_movies_languages['spoken_languages'].apply(literal_eval)
data_movies_languages_exploded = data_movies_languages.explode('spoken_languages')
data_movies_languages_exploded_unnested = pd.json_normalize(data_movies_languages_exploded['spoken_languages']).set_index(data_movies_languages_exploded.pelicula_id)
data_movies_languages_exploded_unnested.reset_index(inplace=True)
data_movies_languages_exploded_unnested.rename(
    columns={'iso_639_1':'spoken_languages_isocode','name' : 'spoken_languages_name'},
    inplace=True
    )
data_movies_languages_exploded_unnested.head(2)

# unir los dataframes
data_movies_normalized = pd.merge(
    pd.merge(pd.merge(pd.merge(data_movies_collection_unnested,
    data_movies_companies_exploded_unnested,
    on=['pelicula_id']),
    data_movies_genres_exploded_unnested,
    on=['pelicula_id']),
    data_movies_countries_exploded_unnested,
    on=['pelicula_id']),
    data_movies_languages_exploded_unnested,
    on=['pelicula_id']
    )
data_movies_normalized.head(2)

data_movies_final=data_movies_normalized[
    data_movies_normalized.columns[~data_movies_normalized.columns.str.contains("_path")] # quitamos las NO contienen "_path"
    ]
data_movies_final.to_csv(os.path.join("2_pipeline","data_movies_normalizada.csv"))
data_movies_final.info()
