==> data_prcssng/03_etl_credits_desanidar_v3.ipynb <==
# 1.3. ETL - Fase 3: Desanidar y normalizar credits

En esta seccion, desinadamos y normalizamos la informacion contenida en `credits.csv`

import os
import numpy as np
import pandas as pd
import json
from ast import literal_eval

# cargamos e inspeccionamos el dataset
data_credits = pd.read_csv(os.path.join('1_data','credits.zip'),compression='zip',low_memory=False).convert_dtypes()
data_credits.rename(columns = {'id':'pelicula_id'}, inplace = True)
data_credits.head()

# nos concentramos en la primera columna
data_credits_cast=pd.DataFrame(data_credits[['cast','pelicula_id']])
data_credits_cast.head()

# problema: Si pasamos la columna a json_normalize nos da un DF vacio:
# esto es por que hay errores de syntaxis en la columna
data_credits_cast_error = pd.json_normalize(data_credits_cast['cast'])
data_credits_cast_error.head()

# pandas.explode() tampoco la entiende: no le hace ninguna transformacion,
data_credits_cast_error2=data_credits_cast.set_index('pelicula_id')
data_credits_cast_error2 = data_credits_cast_error2.explode('cast')
data_credits_cast_error2.head()

# tenemos que corregir la syntaxis de la columna
# primero evaluamos la sintaxis de la columna (tengo entendido que de esta manera corrige cualquier error)

# Esta alternativa no funciona
#data_credits_cast['cast']=data_credits_cast['cast'].str.replace("'", '"', regex=True)
#data_credits_cast['cast']=data_credits_cast['cast'].replace("None", "null")

# la mejor alternativa es primero usar literal_eval, con el fin de asignar la etiqueta de JSON object a la columna
data_credits_cast['cast']=data_credits_cast['cast'].apply(literal_eval)
data_credits_cast.head()

# a pesar de la evaluacion json_normalizar no la entiende
data_credits_cast_error3=data_credits_cast.set_index('pelicula_id')
data_credits_cast_error3_unnested = pd.json_normalize(data_credits_cast_error3['cast']).set_index(data_credits_cast_error3.index)
data_credits_cast_error3_unnested.head(2)

# solucion error3: expandimos la columna para que queden diccionarios anidados (deja de ser una lista de diccionarios)
data_credits_cast_exploded = data_credits_cast.explode('cast', ignore_index=True)
data_credits_cast_exploded.head()

# ahora si se normaliza la columna
data_credits_cast_exploded=data_credits_cast_exploded.set_index('pelicula_id')
data_credits_cast_unnested = pd.json_normalize(data_credits_cast_exploded['cast']).set_index(data_credits_cast_exploded.index)
data_credits_cast_unnested.head()

# inspeccionamos la diferencia entre cast_id y id
print(data_credits_cast_exploded.iloc[0,0])
print(data_credits_cast_exploded.iloc[100,0])
print(data_credits_cast_exploded.iloc[299,0])

print(data_credits_cast_unnested['cast_id'].min(),data_credits_cast_unnested['cast_id'].max())
print(data_credits_cast_unnested['id'].min(),data_credits_cast_unnested['id'].max())
    

# con base en el output, concluimos lo siguiente:
# cast_id: corresponde al id unico para el cast por cada pelicula
# id: corresponde al id unico del cast en toda la base de datos.
# Renombramos
data_credits_cast_unnested.rename(columns = {'id':'unique_id'}, inplace = True)
data_credits_cast_unnested.info()

# si se prefiere el indice como columna
data_credits_cast_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_credits_cast_unnested.head()

# replicamos los pasos anteriores con columna crew
data_credits_crew=pd.DataFrame(data_credits[['crew','pelicula_id']])
data_credits_crew['crew']=data_credits_crew['crew'].apply(literal_eval)
data_credits_crew_exploded = data_credits_crew.explode('crew', ignore_index=True)
data_credits_crew_exploded=data_credits_crew_exploded.set_index('pelicula_id')
data_credits_crew_unnested = pd.json_normalize(data_credits_crew_exploded['crew']).set_index(data_credits_crew_exploded.index)
data_credits_crew_unnested.rename(columns = {'id':'unique_id'}, inplace = True)
data_credits_crew_unnested.reset_index(inplace=True)
data_credits_crew_unnested.head()

# renombramos las columnas de data_credits_crew_unnested: anexamos un prefijo
columns_cast=['character','credit_id','gender','unique_id','name','order','profile_path']
data_credits_cast_unnested.rename(
    columns={c: 'cast_'+c for c in data_credits_cast_unnested.columns if c in columns_cast},
    inplace=True)
data_credits_cast_unnested.head()

# renombramos las columnas de data_credits_crew_unnested
columns_crew=['credit_id','department','gender','unique_id','job','name','profile_path']
data_credits_crew_unnested.rename(
    columns={c: 'crew_'+c for c in data_credits_crew_unnested.columns if c in columns_crew},
    inplace=True)
data_credits_crew_unnested.head()

print("Numero de peliuclas en credits_cast:",data_credits_cast_unnested['pelicula_id'].nunique())
print("Numero de peliuclas en credits_crew:",data_credits_crew_unnested['pelicula_id'].nunique())

# unimos los dataframes; como la disponibilidad de datos es la misma, hacemos un inner join
data_credits_unnested = pd.merge(
    data_credits_cast_unnested,
    data_credits_crew_unnested,
    on=['pelicula_id']
    )
data_credits_unnested.head()

# inspeccionamos las columnas del df final
data_credits_unnested.info()

# seleccionamos unicamente las columnas de interes
columnas_credits = ['pelicula_id','cast_gender','cast_name','cast_order','crew_job','crew_name']
data_credits_unnested_selected=data_credits_unnested[columnas_credits]
data_credits_unnested_selected.info()
data_credits_unnested_selected.head()

# filtramos los valores de interes: primer protagonista y segundo protagonista
# suponemos que cast_order corresponde al orden en que se presenta el elenco de la pelicula
# data_credits_unnested_filtrada["cast_order"].unique()
# con base en lo anterior, suponemos que primer protagonista = 0, y segundo protagonista = 1
# filtramos la base de datos con estos valores
data_credits_unnested_selected_protagonsita_0=data_credits_unnested_selected.loc[
    data_credits_unnested_selected["cast_order"]==0]
data_credits_unnested_selected_protagonsita_1=data_credits_unnested_selected.loc[
    data_credits_unnested_selected["cast_order"]==1]

# confirmamos nuestra suposicion:
print("Cada pelicula_id tiene primer protagonista:",
      all(data_credits_unnested_selected_protagonsita_0.groupby("pelicula_id")['cast_order'].nunique() == 1))
print("Cada pelicula_id tiene segundo protagonista:",
      all(data_credits_unnested_selected_protagonsita_1.groupby("pelicula_id")['cast_order'].nunique() == 1))

# filtramos los valores de interes: equipos de direccion y produccion
#data_credits_unnested_filtrada["crew_department"].unique()
crew_department_filter=['Directing', 'Production'] # seleccion con base en el output anterior

#data_credits_unnested_filtrada["crew_job"].unique()
crew_job_filter=['Director', 'Executive Producer'] # seleccion con base en el output anterior

data_credits_unnested_selected_filtrada=data_credits_unnested_selected.loc[
    (data_credits_unnested_selected["cast_order"].isin([0,1])) &
    #(data_credits_unnested_selected["crew_department"].isin(crew_department_filter)) &
    (data_credits_unnested_selected["crew_job"].isin(crew_job_filter))
]#.dropna()

# renonbramos el data_credits despues de la primera fase de ETL
data_credits_mvp=data_credits_unnested_selected_filtrada # renombramos el df para facilitar la lectura
del data_credits_unnested_selected_filtrada # eliminamos el df previo para liberar espacio en memoria
data_credits_mvp.info()

# Uno de los principales criterios de nuestro sistema de recomendacion el protagonista
# para facilitar la lectura de la columnas, renombramos "prtgnst_" con "prtgnst_"
data_credits_mvp.columns=data_credits_mvp.columns.str.replace('cast_','prtgnst_')
data_credits_mvp.info()

# suponemos que el genero del protagonista esta en formato binario donde 0 = hombre y mujer = 1.
# Confirmamos nuestra suposicion:
print(
    data_credits_mvp.loc[:,['prtgnst_gender','prtgnst_name']][
        data_credits_mvp['prtgnst_gender']==0].head(4)) # hombre: 0
print(data_credits_mvp.loc[:,['prtgnst_gender','prtgnst_name']][
    data_credits_mvp['prtgnst_gender']==1].head(4)) # mujer: 1

#recodificamos la variable prtgnst_gender: asignamos strings para facilitar la lectura.
#preferimos este formato; en dado caso, mas adelante podemos convertir a dummies.
pd.options.mode.chained_assignment = None
data_credits_mvp['prtgnst_gender_strng']= np.where(
    data_credits_mvp.loc[:,['prtgnst_gender']]==1, 'mujer', 'hombre')
data_credits_mvp.drop(columns=['prtgnst_gender'], inplace=True)

# recodificamos la variable prtgnst_order para facilitar lectura: primer protagonsita (1), segundo protagonista (2).
# data_mvp['prtgnst_order'].unique() # array([0., 1.])
data_credits_mvp['prtgnst_nivel']= np.where(
    data_credits_mvp.loc[:,['prtgnst_order']]==0, 1, 2)
data_credits_mvp.drop(columns=['prtgnst_order'], inplace=True)
data_credits_mvp.info()

# Asumimos que cada pelicula tenga un solo Director y productor
# Verificamos nuestra suposicion:
data_credits_mvp_director=data_credits_mvp.loc[
    data_credits_mvp['crew_job'].isin(['Director'])].groupby('pelicula_id').nunique('crew_name')
data_credits_mvp_execproducer=data_credits_mvp.loc[
    data_credits_mvp['crew_job'].isin(['Executive Producer'])].groupby('pelicula_id').nunique('crew_name')
print("Cada pelicula_id tiene un unico Director:",
      all(data_credits_mvp_director['crew_name'] == 1)
      )
print("Cada pelicula_id tiene un unico Productor:",
      all(data_credits_mvp_execproducer['crew_name'] == 1)
      )

# Basado en lo anterior, visualizamos un ejemplo de peliculas con multiples directores.
data_credits_mvp[['pelicula_id','crew_job','crew_name']].loc[
    (data_credits_mvp['pelicula_id']==5) &
    (data_credits_mvp['crew_job'].str.contains('Director',na=False))
    ].drop_duplicates()

# nos concentramos en los valores == Director
data_credits_mvp_2pivot = data_credits_mvp[['pelicula_id','crew_job','crew_name']]
data_credits_mvp_2pivot_director = data_credits_mvp[data_credits_mvp['crew_job'].isin(['Director'])]
data_credits_mvp_director_pivoted=data_credits_mvp_2pivot_director.pivot_table(
    index=['pelicula_id'],
    columns="crew_job",
    values="crew_name",
    #aggfunc='first').reset_index(level='pelicula_id').rename_axis('Index',axis='columns') # toma solo el primer director
    aggfunc=list).reset_index(level='pelicula_id').rename_axis('Index',axis='columns').explode(['Director'], ignore_index=True).drop_duplicates()
data_credits_mvp_director_pivoted.columns=map(str.lower, data_credits_mvp_director_pivoted.columns)
data_credits_mvp_director_pivoted.info()
data_credits_mvp_director_pivoted.head()

# nos concentramos en los valores == Executive Producer
data_credits_mvp_2pivot = data_credits_mvp[['pelicula_id','crew_job','crew_name']]
data_credits_mvp_2pivot_producer = data_credits_mvp[data_credits_mvp['crew_job'].isin(['Executive Producer'])]
data_credits_mvp_producer_pivoted=data_credits_mvp_2pivot_producer.pivot_table(
    index=['pelicula_id'],
    columns="crew_job",
    values="crew_name",
    #aggfunc='first').reset_index(level='pelicula_id').rename_axis('Index',axis='columns') # toma solo el primer productor
    aggfunc=list).reset_index(level='pelicula_id').rename_axis('Index',axis='columns').explode(['Executive Producer'], ignore_index=True).drop_duplicates()
data_credits_mvp_producer_pivoted.columns=map(str.lower, data_credits_mvp_producer_pivoted.columns.str.replace(' ','_'))
data_credits_mvp_producer_pivoted.info()
data_credits_mvp_producer_pivoted.head()

# Evaluamos cual de las dos dataframes contiene el mayor numero de registros;
# El df con el mayor numero de registros sera nuestra df de referencia en el merge
print("Disponibilidad de registros para director:",data_credits_mvp_director_pivoted['pelicula_id'].nunique())
print("Disponibilidad de registros para productor:",data_credits_mvp_producer_pivoted['pelicula_id'].nunique())

# unimos los dfs
data_credits_mvp_crew_pivoted=pd.merge(
    data_credits_mvp_director_pivoted, # tomamos como referencia la base de datos con mayor numero de registros
    data_credits_mvp_producer_pivoted,
    how='left', # tomamos como referencia la base de datos con mayor numero de registros
    on='pelicula_id')
data_credits_mvp_crew_pivoted.info()
data_credits_mvp_crew_pivoted.head()

# Confirmamos que df con el mayor numero de registros es nuestra df principal: data_credits_mvp
print("Disponibilidad de registros por peliculas en df principal:",data_credits_mvp['pelicula_id'].nunique()) # df principal
print("Disponibilidad de registros por peliculas en df crew normalizada:",data_credits_mvp_crew_pivoted['pelicula_id'].nunique())

# primero, removemos las columnas en el df principal para evitar redundancia
data_credits_mvp_2joincrew=data_credits_mvp.drop(columns=['crew_job','crew_name']).drop_duplicates()

# segundo, unimos las df con el df global
data_credits_mvp_final=pd.merge(
    data_credits_mvp_2joincrew, # tomamos como punto de referencia nuestra df principal: data_credits_mvp
    data_credits_mvp_crew_pivoted,
    how='left',
    on='pelicula_id')
data_credits_mvp_final.to_csv(os.path.join("2_pipeline","credits_normalizada.csv"))
data_credits_mvp_final.info()
data_credits_mvp_final.head(2)
