==> data_prcssng/08_api_consult.ipynb <==
# API Interaction

## Objectives
1. Demonstrate how to interact with the IMDB-based API generated earlier.
2. Retrieve and process data about directors and movies.
3. Show how to handle nested JSON responses and convert them to pandas DataFrames.

## Key Techniques Demonstrated
- API requests using the `requests` library
- JSON data processing
- Handling nested data structures
- Pandas DataFrame manipulation (json_normalize, merging, type conversion)

## Contents
1. Import packages
2. Test the API connection
3. Data retrieval and manipulation

----
## 1. Import packages

import requests
import pandas as pd

----
## 2. Test API connection

# Define the API endpoint URL for the Recommendation system based on IMDB data
api_imdb = "https://henry-mlops-imdb.onrender.com/"

# Send a GET request to the IMDB API endpoint
response_api = requests.get(api_imdb)

# Print the raw text response from the API
print(response_api.text)

----
## 3. Data retrieval and manipulation

### 3.1. Director Information Retrieval
First, we fetch data about a specific director (Woody Allen).


# Define the API URL with placeholder for director name
url_imdb_director = "https://henry-mlops-imdb.onrender.com/director_exitoso/{}"

# Define the director name
query_director = "Woody Allen"

# Format the API URL with the director name
api_imdb_director = url_imdb_director.format(query_director)

# Send the GET request using requests
response_api_director = requests.get(api_imdb_director)

data_director_query = pd.DataFrame(response_api_director.json())
data_director_query.head(3)

Second, we process the nested JSON response.
Third, we create a structured DataFrame with director and movie information

# Set the index of the DataFrame to the 'director' column
data_director_query_nested = data_director_query.set_index("director")

# Unnest the nested list using pandas.json_normalize and set the index to the original index
data_director_query_unnested = pd.json_normalize(
    data_director_query_nested["peliculas"]).set_index(data_director_query_nested.index)

# Merge the 'director' information DataFrame (excluding the 'peliculas' column and duplicates)
# with the unnested movie DataFrame on the 'director' index
data_director_expanded = pd.merge(
    data_director_query.drop(columns=["peliculas"]).drop_duplicates().set_index("director"),
    data_director_query_unnested,
    left_index=True, right_index=True
# Reset the index to create a new column for the index and convert data types
).reset_index().convert_dtypes()

data_director_expanded.info()
data_director_expanded.head()

### 3.2. Movie Recommendations
We retrieve movie recommendations based on a given movie title, and display the recommendations in a pandas DataFrame.

# Define the API URL with placeholder for director name
url_imdb_recommend = "https://henry-mlops-imdb.onrender.com/recomendacion_pelicula/{}"

# Define the director name
query_movie = "Top Gun"

# Format the API URL with the director name
api_imdb_movie = url_imdb_recommend.format(query_movie)

# Send the GET request using requests
response_api_movie = requests.get(api_imdb_movie)

data_movie_query = pd.DataFrame(response_api_movie.json())
data_movie_query.head(5)
